backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
